requirements.txt
torch~=2.6.0
torchvision~=0.21.0
timm~=0.9.2
pycocotools~=2.0.8
matplotlib~=3.10.0rc1
numpy~=2.1.2
opencv-python~=4.11.0.86
scikit-learn~=1.5.2
Pillow~=11.0.0
tqdm~=4.67.1
pandas
argparse~=1.4.0
requests~=2.32.3
Flask~=3.1.0
dotenv